+++ engine/language-server/src/main/scala/org/enso/languageserver/filemanager/FileEventWatcher.scala
@@ -23,48 +23,24 @@ final class FileEventWatcher(
     .listener(this)
     .build()
 
+  /**
+    * Start the watcher.
+    */
   def start(): Unit =
     watcher.watch()
 
+  /**
+    * Stop the watcher.
+    */
   def stop(): Unit =
     watcher.close()
 
+  /**
+    * A callback executed by [[DirectoryWatcher]] on file system event.
+    */
   override def onEvent(event: DirectoryChangeEvent): Unit = {
     WatcherEvent
       .from(event)
       .foreach(callback)
-  }
-}
-
-object FileEventWatcherApi {
-
-  sealed trait EventType
-
-  object EventType {
-
-    def from(eventType: DirectoryChangeEvent.EventType): Option[EventType] =
-      eventType match {
-        case DirectoryChangeEvent.EventType.CREATE   => Some(EventTypeCreate)
-        case DirectoryChangeEvent.EventType.MODIFY   => Some(EventTypeModify)
-        case DirectoryChangeEvent.EventType.DELETE   => Some(EventTypeDelete)
-        case DirectoryChangeEvent.EventType.OVERFLOW => None
-      }
-  }
-
-  case object EventTypeCreate extends EventType
-
-  case object EventTypeModify extends EventType
-
-  case object EventTypeDelete extends EventType
-
-  case class WatcherEvent(path: Path, eventType: EventType)
-
-  object WatcherEvent {
-
-    def from(event: DirectoryChangeEvent): Option[WatcherEvent] =
-      EventType
-        .from(event.eventType())
-        .map(WatcherEvent(event.path(), _))
-
   }
 }
